+++ MercadoLibre/Common/API/Services/Items/Errors/ItemServiceErrors.swift
@@ -9,6 +9,7 @@ import Foundation
 
 enum ItemServiceErrors: Error {
     case unableToLoadMock
+    case productIdIsEmpty
     case itemNotFoundInResults
     case productTranslatorNotFound
     case simulatedError