+++ MercadoLibre/Helpers/Theme/MeliTheme.swift
@@ -1,0 +1,15 @@
+//
+//  MeliTheme.swift
+//  MercadoLibre
+//
+//  Created by John Edwin Guerrero Ayala on 19/04/21.
+//
+
+import Foundation
+import UIKit
+
+enum MeliTheme {
+    enum Colors {
+        static let backgroundColor = UIColor(named: "MeliBackground")
+    }
+}