+++ MercadoLibre/Modules/Home/CellPrototype/HomeTableViewCell.swift
@@ -6,13 +6,15 @@
 //
 
 import UIKit
+import Kingfisher
 
 class HomeTableViewCell: UITableViewCell {
 
     // MARK: - IBOutlets -
-    
+
+    @IBOutlet weak var thumbnailImageView: UIImageView!
     @IBOutlet weak var titleLabel: UILabel!
-    
+
     override func awakeFromNib() {
         super.awakeFromNib()
         // Initialization code
@@ -23,39 +25,65 @@ class HomeTableViewCell: UITableViewCell {
 
         // Configure the view for the selected state
     }
-    
+
     func configure(with cellViewModel: CellViewModel) {
         guard let productCellViewModel = cellViewModel as? ProductCellViewModel else {
             return
         }
         titleLabel.text = productCellViewModel.title
+        updateThumbnailImage(urlString: productCellViewModel.thumbnailUrl)
     }
-    
+
+    func updateThumbnailImage(urlString: String) {
+        let url = URL(string: urlString)
+        let placeholderImage = #imageLiteral(resourceName: "placeHolder")
+        thumbnailImageView.kf.setImage(
+            with: url,
+            placeholder: placeholderImage,
+            options: [
+                .loadDiskFileSynchronously,
+                .cacheOriginalImage,
+                .transition(.fade(0.25))
+            ]
+        )
+    }
+
 }
 
 extension HomeTableViewCell: TableViewCellCustomizable {
-    
+
     static var cellId: String {
         return "HomeViewCell"
     }
-    
+
     static var bundle: Bundle {
         return Bundle(for: HomeTableViewCell.self)
     }
-    
+
     static var nib: UINib {
         return UINib(nibName: String(describing: self), bundle: HomeTableViewCell.bundle)
     }
-    
+
     static func registerCellPrototypes(tableView: UITableView) {
         tableView.register(HomeTableViewCell.nib, forCellReuseIdentifier: HomeTableViewCell.cellId)
     }
-    
-    static func dequeue(from tableView: UITableView, for indexPath: IndexPath, with cellViewModel: CellViewModel) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: HomeTableViewCell.cellId, for: indexPath) as! HomeTableViewCell
+
+    static func dequeue(
+        from tableView: UITableView,
+        for indexPath: IndexPath,
+        with cellViewModel: CellViewModel
+    ) -> UITableViewCell {
+        guard let cell = tableView
+                .dequeueReusableCell(
+                    withIdentifier: HomeTableViewCell.cellId,
+                    for: indexPath
+                ) as? HomeTableViewCell else {
+            return UITableViewCell()
+        }
+
         cell.configure(with: cellViewModel)
 
         return cell
     }
-    
+
 }