+++ MercadoLibre/Modules/Home/HomeView.swift
@@ -8,9 +8,10 @@
 import Foundation
 
 protocol HomeView: AnyObject {
+    var searchTerm: String { get set }
     func startLoading()
     func finishLoading()
-    func setResults(with products: [Product])
+    func reloadAndShowTableView()
     func setEmptyView()
     func setErrorView()
 }