+++ MercadoLibre/Modules/Home/HomeViewModel.swift
@@ -8,103 +8,71 @@
 import Foundation
 import UIKit
 
-struct HomeStandByState: HomeViewState {
-    
-}
+final class HomeViewModel {
 
-struct HomeSearchState: HomeViewState {
-    var context: HomeViewModel
-    
-    func search(by term: String) {
-        guard !context.searchTerm.isEmpty else {
-            context.state = HomeStandByState()
-            context.homeView.setEmptyView()
-            context.homeView.finishLoading()
-            return
-        }
-        context.state = HomeSearchingState(context: context)
-        context.state.search(by: term)
-    }
-}
-
-struct HomeSearchingState: HomeViewState {
-    var context: HomeViewModel
-    
-    func search(by term: String) {
-        context.homeView.startLoading()
-        context.itemService.items(byTerm: term) { (responseProducts, responseError) in
-            context.state = HomeProcessingState(context: context, products: responseProducts, error: responseError)
-            context.currentState.handleResponse()
-        }
-    }
-}
-
-struct HomeProcessingState: HomeViewState {
-    var context: HomeViewModel
-    var products: [Product]
-    var error: Error?
-    
-    func handleResponse() {
-        defer {
-            context.state = HomeStandByState()
-            context.homeView.finishLoading()
-        }
-        
-        guard error == nil else {
-            context.homeView.setErrorView()
-            return
-        }
-        
-        guard !products.isEmpty else {
-            context.homeView.setEmptyView()
-            return
-        }
-        context.homeView.setResults(with: products)
-    }
-}
-
-class HomeViewModel {
-    
     // MARK: - Public properties -
-    
+
+    var isPerformingSearch: Bool = false
+
     var homeView: HomeView {
         return view
     }
-    
+
     var currentState: HomeViewState {
-        return state
-    }
-    
-    var itemService: ItemsService {
-        return service
-    }
-    
-    var searchTerm: String = "" {
-        didSet {
-            startSearching()
+        get {
+            return state
+        }
+        set {
+            state = newValue
         }
     }
-    
+
+    var itemService: ItemsService? {
+        return service
+    }
+
+    var products: [Product] = []
+
+    var productsCount: Int {
+        return products.count
+    }
+
+    var searchTerm: String {
+        return view.searchTerm
+    }
+
     // MARK: - Private properties -
-    
+
     private unowned let view: HomeView
-    fileprivate var state: HomeViewState = HomeStandByState()
-    fileprivate var service: ItemsService
-    
+    var state: HomeViewState = HomeStandByState()
+    fileprivate var service: ItemsService!
+
     init(view: HomeView, service: ItemsService) {
         self.view = view
         self.service = service
     }
-    
+
     convenience init(view: HomeView) {
         let alamofireService = AlamofireItemService()
         self.init(view: view, service: alamofireService)
     }
-    
+
 }
 
 extension HomeViewModel {
+    func setResults(with products: [Product]) {
+        self.products = products
+        homeView.reloadAndShowTableView()
+    }
+
+    func product(at index: Int) -> Product {
+        return products[index]
+    }
+
     func startSearching() {
+        if isPerformingSearch {
+            itemService?.cancelAllRequest()
+        }
         state = HomeSearchState(context: self)
         state.search(by: searchTerm)
     }