+++ MercadoLibreTests/Model/ProductTests.swift
@@ -10,11 +10,11 @@ import XCTest
 
 class ProductTests: XCTestCase {
 
-    var sut: Item!
+    var sut: Product!
     
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
-        sut = Product(id: "id", siteId: "siteId", title: "Product title", price: 10000)
+        sut = Product(id: "id", siteId: "siteId", title: "Product title", price: 10000, thumbnailUrl: "thumbnail")
     }
 
     override func tearDownWithError() throws {